requirements.txt
annotated-types==0.7.0
anyio==4.6.0
bottle==0.13.1
bottle-websocket==0.2.9
certifi==2024.8.30
cffi==1.17.1
colorama==0.4.6
distro==1.9.0
Eel==0.17.0
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
openai==1.51.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.4
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
whichcraft==0.6.1
zope.event==5.0
zope.interface==7.0.3